requirements.txt
# FastAPI 관련 패키지
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
aiofiles==23.2.1
websockets==12.0
pydantic==2.4.2
pydantic-settings==2.0.3

# 컴퓨터 비전 및 AI 패키지
opencv-python-headless==4.8.1.78
numpy==1.24.3
Pillow==10.0.1
ultralytics>=8.1.0,<9.0.0
torch==2.1.0
torchvision==0.16.0

# Redis 관련 패키지
redis[hiredis]==5.0.1

# OpenStack Swift 관련 패키지
python-swiftclient==4.4.0
python-keystoneclient==5.1.0

# 기존 패키지들
requests>=2.28.0
mysql-connector-python>=8.0.0
PyMySQL>=1.0.0
shapely>=2.0.0
python-dotenv>=1.0.0
pandas>=1.5.0
